17 April - Code to solve wave + blowup equations
Nice features. In case we have only wave the wave travels, while for the case where we turn on the blowup term, for large enough amplitude it blowup. We also found that the reason it doesn't work well for the cos function is that  cos is not localized enough and we cant see the wave moving nicely.